Lots of progress on operation chainning
I figured out the logic to chain operators, use decimals, and clear all operations. I decided to get rid of the feature where the calculator displays the result as you type in operations instead of you having to press the equals button because it was too complicated. I now need to handle negative integers, cases where the user tries to enter multiple decimals for one number, and the "C" button.