Update calculator.js
I added negative number functionality to the calculator as well as the ability to perform an operation on a pair of numbers then display the result in the upper part of the screen so it can be used in the next operation. I now need to work a little more on the aesthetics of the calculator.